Move action_menu to frontend and port to TS
