+++ app/assets/javascripts/application.js.erb
@@ -28,7 +28,6 @@
 
 //= require lib/jquery.colorcontrast
 
-//= require action_menu
 //= require settings
 //= require openproject_plugins
